JAMES-1818 Move Attachment and AttachmentId to API
