+++ mailbox/api/src/test/java/org/apache/james/mailbox/model/AttachmentIdTest.java
@@ -17,10 +17,11 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailbox.store.mail.model;
+package org.apache.james.mailbox.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.james.mailbox.model.AttachmentId;
 import org.junit.Test;
 
 public class AttachmentIdTest {
